+++ node/CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.  The format
 * Add `tcp_connect_timeout`, `setup_timeout`, `tcp_connect_attempts`, `tcp_connect_base_backoff`, `significant_error_backoff`, `permanent_error_backoff`, `successful_reconnect_delay`, `flaky_connection_threshold`, `max_incoming_connections` and `max_outgoing_connections` to the `network.conman` section in the config.
 * `use_validator_broadcast` can now be configured to control the node's broadcast behavior.
 * `use_mixed_gossip` can now be configured to enable or disable the node's gossip peer selection.
+* Add `net_gossip_requests`, `net_overflow_buffer_count` and `net_overflow_buffer_bytes` metrics.
 
 ### Changed
 * The node's connection model has changed, now only establishing a single connection per peer. The direction of the connection is chosen based on the randomly generated `NodeID`s.